autofix-robotai/requirements.txt
crewai==0.35.0
chromadb>=0.4.24,<0.5.0
sentence-transformers==3.0.1
robotframework==6.1.1
Appium-Python-Client==3.1.0
ollama==0.2.1
GitPython==3.1.43
beautifulsoup4==4.12.3
lxml==5.2.2
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
click==8.1.7
setuptools==70.0.0
